requirements.txt
aldryn-boilerplates==0.8.0
asgiref==3.2.7
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
disqus-python==0.4.2
dj-database-url==0.5.0
Django==2.2
django-absolute==0.3
django-admin-jqueryui==1.8.22
django-admin-sortable==2.2.3
django-appconf==1.0.4
django-classy-tags==1.0.0
django-cms==3.7.2
django-connected==0.1.5
django-crispy-forms==1.9.1
django-email-extras==0.3.4
django-emailit==0.2.4
django-entangled==0.3
django-filer==1.7.1
django-forms-builder==0.14.0
django-formtools==2.2
django-inline-ordering==1.0.2
django-ipware==2.1.0
django-js-asset==1.2.2
django-localflavor==3.0.1
django-meta==1.6.1
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-polymorphic==2.1.2
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-responsive2==0.1.3
django-sekizai==1.1.0
django-simple-captcha==0.5.12
django-simple-math-captcha==1.0.9
django-sizefield==1.0.0
django-tablib==3.2
django-taggit==1.2.0
django-taggit-autosuggest==0.3.8
django-treebeard==4.3.1
djangocms-accordion==1.0.4
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.6.0
djangocms-cascade==1.3.1
djangocms-column==1.10.0
djangocms-disqus==0.1.4
djangocms-file==2.4.0
djangocms-forms==0.2.5
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-inherit==0.2.2
djangocms-instagram==0.1.1
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-owl==0.1.11
djangocms-page-meta==0.10.0
djangocms-page-tags==0.8.0
djangocms-picture==2.4.0
djangocms-responsive-wrapper==0.2.1
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-table==1.2
djangocms-text-ckeditor==3.9.0
djangocms-timer==0.2.0
djangocms-twitter2==0.1.1
djangocms-video==2.3.0
djangocms-youtube==0.2.1
easy-thumbnails==2.7
et-xmlfile==1.0.1
future==0.15.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gunicorn==20.0.4
hashids==1.2.0
html5lib==1.0.1
httplib2==0.18.1
idna==2.9
isodate==0.6.0
jdcal==1.4.1
jsonfield==3.1.0
lxml==4.5.1
oauthlib==3.1.0
openpyxl==2.4.9
phonenumbers==8.12.4
Pillow==7.1.2
premailer==3.7.0
protobuf==3.12.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrate==0.5a9
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-gnupg==0.4.6
python-instagram==1.3.2
python-stdnum==1.13
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
simplejson==3.17.0
six==1.14.0
South==1.0.2
sphinx-me==0.3
sqlparse==0.3.1
tablib==2.0.0
tweepy==3.8.0
tzlocal==2.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
YURL==1.0.0
